Tamaño de cama parametrizado

+++ CONSTANTS.h
@@ -1,8 +1,8 @@
 #ifndef CONSTANTS_H
 #define CONSTANTS_H
 
-constexpr int TableX = 220;
-constexpr int TableY = 220;
+constexpr double TableX = 220;
+constexpr double TableY = 220;
 
 constexpr int maxDensity = 10;
 #endif // CONSTANTS_H
